代码调整 - ScriptClient
